requirements.txt
# Snowflake Cortex GenAI Pipeline Dependencies

# Core Snowflake packages
snowflake-connector-python>=3.4.0
snowflake-snowpark-python>=1.8.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Document processing  
PyPDF2>=3.0.1
python-docx>=0.8.11

# Configuration and environment
python-dotenv>=1.0.0
pyyaml>=6.0

# Web dashboard
streamlit>=1.28.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development and testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Jupyter support (optional)
jupyter>=1.0.0
ipykernel>=6.25.0